pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tval"
version = "0.1.0"
description = "Table data validator for pre-analysis data validation"
requires-python = ">=3.9"
dependencies = [
    "duckdb",
    "pydantic",
    "pyyaml",
    "jinja2",
    "chardet",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "mypy",
    "types-pyyaml",
    "types-chardet",
]

[project.scripts]
tval = "tval.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/tval"]

[tool.ruff]
target-version = "py39"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]   # pycodestyle, pyflakes, isort

[tool.mypy]
python_version = "3.9"
strict = true
